+++ adapters/dimmable_bulb_adapter.py
@@ -1,15 +1,16 @@
-import Domoticz
 import json
 from adapters.base_adapter import Adapter
 from devices.switch.dimmer_switch import DimmerSwitch
 
 
-class TradfriDimmableBulb(Adapter):
+class DimmableBulbAdapter(Adapter):
     def __init__(self, devices):
         super().__init__(devices)
         self.devices.append(DimmerSwitch(devices, 'light', 'value'))
 
     def convert_message(self, message):
+        message = super().convert_message(message)
+
         if 'state' in message.raw and message.raw['state'].upper() == 'OFF':
             message.raw['value'] = 0
         elif 'brightness' in message.raw:
@@ -33,6 +34,6 @@ class TradfriDimmableBulb(Adapter):
                 'topic': device_data['friendly_name'] + '/set',
                 'payload': json.dumps({
                     "state": "ON",
-                    "brightness": int(level*255/100)
+                    "brightness": int(level * 255 / 100)
                 })
             }